+++ src/dimmer_inline_asm.h
@@ -1,20 +1,28 @@
 // AUTO GENERATED FILE - DO NOT MODIFY
 // MCU: ATmega328P
+// zero crossing pin=PIND mask=0x08/b00001000
+// port=PORTB mask=0x02 pin=1 mask=b00000010/0x02
 // port=PORTD mask=0x40 pin=6 mask=b01000000/0x40
 
 #pragma once
 #include <avr/io.h>
 
+#define DIMMER_SFR_ZC_STATE (PIND & 0x08)
+#define DIMMER_SFR_ZC_IS_SET asm volatile ("sbis %0, %1" :: "I" ( _SFR_IO_ADDR(PIND)), "I" (3));
+#define DIMMER_SFR_ZC_IS_CLR asm volatile ("sbic %0, %1" :: "I" ( _SFR_IO_ADDR(PIND)), "I" (3));
+#define DIMMER_SFR_ZC_JMP_IF_SET(label) DIMMER_SFR_ZC_IS_CLR asm volatile("rjmp " # label);
+#define DIMMER_SFR_ZC_JMP_IF_CLR(label) DIMMER_SFR_ZC_IS_SET asm volatile("rjmp " # label);
+
 // all ports are read first, modified and written back to minimize the delay between toggling different ports (3 cycles, 187.5ns per port @16MHz)
 
-#define DIMMER_SFR_ENABLE_ALL_CHANNELS() { PORTD |= 0x40; }
-#define DIMMER_SFR_DISABLE_ALL_CHANNELS() { PORTD &= ~0x40; }
-#define DIMMER_SFR_CHANNELS_SET_BITS(mask) { PORTD |= mask; }
-#define DIMMER_SFR_CHANNELS_CLR_BITS(mask) { PORTD &= ~mask; }
+#define DIMMER_SFR_ENABLE_ALL_CHANNELS() { uint8_t tmp[2] = { (uint8_t)(PORTB | 0x02), (uint8_t)(PORTD | 0x40) }; PORTB = tmp[0]; PORTD = tmp[1]; }
+#define DIMMER_SFR_DISABLE_ALL_CHANNELS() { uint8_t tmp[2] = { (uint8_t)(PORTB & ~0x02), (uint8_t)(PORTD & ~0x40) }; PORTB = tmp[0]; PORTD = tmp[1]; }
+#define DIMMER_SFR_CHANNELS_SET_BITS(mask) { uint8_t tmp[2] = { (uint8_t)(PORTB | mask[0]), (uint8_t)(PORTD | mask[1]) }; PORTB = tmp[0]; PORTD = tmp[1]; }
+#define DIMMER_SFR_CHANNELS_CLR_BITS(mask) { uint8_t tmp[2] = { (uint8_t)(PORTB & ~mask[0]), (uint8_t)(PORTD & ~mask[1]) }; PORTB = tmp[0]; PORTD = tmp[1]; }
 
-#define DIMMER_SFR_CHANNELS_ENABLE(channel) asm volatile ("sbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTD)), "I" (6))
-#define DIMMER_SFR_CHANNELS_DISABLE(channel) asm volatile ("cbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTD)), "I" (6))
+#define DIMMER_SFR_CHANNELS_ENABLE(channel) switch(channel) { default: asm volatile ("sbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTB)), "I" (1)); break; case 1: asm volatile ("sbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTD)), "I" (6)); }
+#define DIMMER_SFR_CHANNELS_DISABLE(channel) switch(channel) { default: asm volatile ("cbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTB)), "I" (1)); break; case 1: asm volatile ("cbi %0, %1" :: "I" ( _SFR_IO_ADDR(PORTD)), "I" (6)); }
 
 // verify signature during compliation
-static constexpr uint8_t kInlineAssemblerSignature[] = { 0x1e, 0x95, 0x0f, 0x06 }; // MCU ATmega328P (1e-95-0f)
+static constexpr uint8_t kInlineAssemblerSignature[] = { 0x1e, 0x95, 0x0f, 0x09, 0x06 }; // MCU ATmega328P (1e-95-0f)